% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{fetch_test4_sample_data}
\alias{fetch_test4_sample_data}
\title{fetch the Test4 sample data}
\usage{
fetch_test4_sample_data()
}
\value{
a named list of data frames
}
\description{
fetch the Test4 sample data
}
\author{
karl
}

