% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tm_etl.R
\name{run_tm_etl}
\alias{run_tm_etl}
\title{run the tMDataLoader tool}
\usage{
run_tm_etl(data_dfs, map_df, etl_path, ...)
}
\arguments{
\item{data_dfs}{the data to upload as a (list of) data frame(s)}

\item{etl_path}{the path in etl tree, as a standard path string}

\item{...}{additional arguments to \code{run_etl_command}}
}
\value{
the upload output log and summary statistics as a list
}
\description{
Will reformat the data and generate automatically the Mapping file using
the supplied mapper map_df
}
\author{
karl
}

