% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tmdataloader.R
\name{delete_study_by_id}
\alias{delete_study_by_id}
\title{delete a study by id using the tMDataLoader tool}
\usage{
delete_study_by_id(id, ...)
}
\arguments{
\item{id}{the id of the study to delete}

\item{...}{passed to run_etl_command}
}
\value{
the command output
}
\description{
delete a study by id using the tMDataLoader tool
}
\author{
karl
}

