% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run_tm_etl}
\alias{run_tm_etl}
\title{run the tMDataLoader tool}
\usage{
run_tm_etl(data_dfs, mapping_df, host = "localhost", port = 5432,
  java = "java", jar = find_extdata_file("tm_etl.jar"),
  config = find_extdata_file("Config.groovy.brew"), dir = NULL)
}
\arguments{
\item{data_dfs}{the data to upload}

\item{mapping_df}{the corresponding mapping}

\item{host}{the transmart DB host}

\item{port}{the transmart DB port}

\item{java}{the java executable to use}

\item{jar}{the tMDataLoader jar file}

\item{config}{the tMDataLoader config file template}

\item{dir}{the directory in which to setup the data files
                                            to upload. If not given will occur in a temp dir}
}
\description{
run the tMDataLoader tool
}
\author{
karl
}

