% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tmdataloader.R
\name{run_tm_etl}
\alias{run_tm_etl}
\title{run the tMDataLoader tool}
\usage{
run_tm_etl(data_dfs, map_df, etl_path, ...)
}
\arguments{
\item{data_dfs}{the data to upload}

\item{map_df}{the corresponding mapping}

\item{etl_path}{the path in etl tree, as a standard path string}

\item{...}{additional arguments to \code{run_etl_command}}
}
\value{
the upload summary statistics as a data frame, or NULL
}
\description{
run the tMDataLoader tool
}
\author{
karl
}

